Flyttet vedtaksbelop til egen klasse
